add cuda bitonic sort

+++ sorting/cuda/array_utils.cpp
@@ -5,15 +5,15 @@
 #include <time.h>
 
 template<typename T>
-void Generator<T>::random(std::vector<T> &a, u_int32_t size) {
+void Generator<T>::random(std::vector<T> &a, uint64_t size) {
 	srand (time(NULL));
-	for (int i=0;i<size;i++)
+	for (uint64_t i=0;i<size;i++)
 		a[i] = (T)rand() % (MAX - MIN) + MIN;
 }
 
 template<typename T>
-void Generator<T>::increasing(std::vector<T> &a, u_int32_t size) {
+void Generator<T>::increasing(std::vector<T> &a, uint64_t size) {
 	srand (time(NULL));
-	for (int i=0;i<size;i++)
+	for (uint64_t i=0;i<size;i++)
 		a[i] = rand() % (MAX - MIN) + MIN;
 }
